+++ tests/fixtures/source/MainLoop.xs
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
+// SPDX-License-Identifier: Apache-2.0
+
+fn main() -> Long {
+    value: Long = 0;
+    loop {
+        value += 1;
+        if (value == 7) {
+            break;
+        }
+    }
+    return value;
+}